Add messages to breakdown and download about filters
